finished first stage of rewrite

+++ packages/states/post/confirmpostdeletionstate.py
@@ -18,7 +18,7 @@ class ConfirmPostDeletionState(SelectDraftUpdateState):
     def welcome_message(self):
         message = "It seems the post you selected no longer exists..."
 
-        post = self.context.get_post(self.post_id)
+        post = self.context.persistence.get_post(self.post_id)
         if post is not None:
             message = "Do you <b>really</b> want to <b>delete</b> post <b>" + post.title + "</b>?"
 
@@ -31,7 +31,7 @@ class ConfirmPostDeletionState(SelectDraftUpdateState):
         reply_options = [{"text": "<< drafts", "callback_data": "/deletepost"}, []]
 
         # if post still exists, show button to confirm final deletion
-        post = self.context.get_post(self.post_id)
+        post = self.context.persistence.get_post(self.post_id)
         if post is not None:
             reply_options.append({"text": "YES, delete", "callback_data": "/confirmpostdeletion /confirm"})
             reply_options.append([])
@@ -51,7 +51,7 @@ class ConfirmPostDeletionState(SelectDraftUpdateState):
             # confirmed post deletion
             if command_array[1] == "/confirm":
 
-                deleted_post = self.context.delete_post(self.post_id)
+                deleted_post = self.context.delete_post(self.post_id) if self.context.unpublish(self.post_id) else None
 
                 # post removal successful
                 if deleted_post is not None:
@@ -65,8 +65,9 @@ class ConfirmPostDeletionState(SelectDraftUpdateState):
                                               , parse_mode=ParseMode.HTML.value)
 
                 # show remaining posts for deletion
-                user_posts = self.context.get_user_posts(user_id=user_id, status=PostState.PUBLISHED)
-                if len(user_posts) > 0:
+                user_posts = self.context.persistence.get_posts(user_id=user_id, status=PostState.PUBLISHED)
+                if len([post for post in user_posts if post.status == PostState.PUBLISHED]) \
+                        > len([draft for draft in user_posts if draft.status == PostState.DRAFT and draft.original_post is not None]):
                     from packages.states.post.deletepoststate import DeletePostState
                     next_state = DeletePostState(self.context, user_id, chat_id=chat_id)
                 # no remaining posts -> automatically go back to main menu
